backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
openai>=1.10.0
pymongo==4.0.1
PyJWT==2.3.0
Werkzeug==2.0.1
langchain==0.1.0
langchain-openai==0.0.5
python-i18n==0.3.9
langdetect==1.0.9 